Apply Rob's HTML and CSS for the new UI
Copy Rob's markup for the fields,
and the CSS file.
